refactor(#58) :: home 폴더 구조 리팩토링
