+++ src/features/Home/Category/index.tsx
@@ -1,5 +1,5 @@
-import * as s from './ui/style.css';
-import { BtnCategoryPlus, BtnCategoryItem } from './ui';
+import * as s from './style.css';
+import { BtnCategoryPlus, BtnCategoryItem } from '../ui';
 
 const Category = () => {
   return (
