--- src/features/Home/Todo/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as BtnToDoNormal } from './BtnToDoNormal';
-export { default as BtnToDoChecked } from './BtnToDoChecked';
