+++ src/features/Home/ui/index.ts
@@ -1,2 +1,4 @@
+export { default as BtnToDoNormal } from './BtnToDoNormal';
+export { default as BtnToDoChecked } from './BtnToDoChecked';
 export { default as BtnCategoryPlus } from './BtnCategoryPlus';
 export { default as BtnCategoryItem } from './BtnCategoryItem';
